requirements.txt
requests
loguru
pycryptodome
opencv-python
fire
flask
pyexecjs
supervisor
fastapi[all]